requirements.txt
openai>=1.30.0          # OpenRouter uses the OpenAI-compatible API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# PDF parsing (pdfplumber > pypdf for tables + complex layouts)
pdfplumber>=0.10.0
pypdf>=4.0.0          # fallback

# Embeddings
sentence-transformers>=2.6.0

# Vector store
qdrant-client>=1.8.0

# Lexical search
rank-bm25>=0.2.2

# Tokenisation
tiktoken>=0.6.0

# ColBERT
ragatouille>=0.0.7
langchain>=0.1.0      # ragatouille dependency

# Clustering (RAPTOR)
scikit-learn>=1.4.0

# Rate limiting
slowapi>=0.1.9

# HTTP
httpx>=0.27.0

numpy>=1.24.0
